Extract NIMs and MDEs
